# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Boilerplate V1",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Lista de User",
                "description": "Devuelve una lista de User",
                "parameters": [
                    {
                        "name": "paginable",
                        "in": "query",
                        "description": "Define si el resultado estará paginado",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Define la cantidad de resultados por página",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "relations",
                        "in": "query",
                        "description": "Define la un arreglo de las relaciones a obtener",
                        "required": false,
                        "style": "deepObject",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra los User.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Crear un User",
                "description": "Devuelve el User creado",
                "requestBody": {
                    "request": "User",
                    "description": "Objeto del User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve el User creado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Falta algún parámetro."
                    },
                    "400": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Obtener un User",
                "description": "Obtener un User por su ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "relations",
                        "in": "query",
                        "description": "Define un arreglo de las relaciones a obtener",
                        "required": false,
                        "style": "deepObject",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra la respuesta.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El registro no se encuentra en la base de datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Elimina un User",
                "description": "Elimina un User por su ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un código 200."
                    },
                    "400": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El registro no se encuentra en la base de datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Actualizar un User",
                "description": "Permite editar un User mediante su ID",
                "requestBody": {
                    "request": "User",
                    "description": "Objeto del User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve el User actualizado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Falta algún parámetro."
                    },
                    "400": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/users/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inicio de sesión",
                "description": "Requiere el email del usuario y su contraseña para iniciar sesión",
                "requestBody": {
                    "request": "Login",
                    "description": "Objeto del inicio de sesión",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "El inicio de sesión fue exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Algún campo no es correcto.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "created_at": {
                        "description": "Fecha de creación del User",
                        "type": "string",
                        "example": "2020-02-06 11:26:24"
                    },
                    "updated_at": {
                        "description": "Fecha de edición del User",
                        "type": "string",
                        "example": "2020-02-06 11:26:24"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "title": "Login Request",
                "description": "Login Request",
                "properties": {
                    "email": {
                        "description": "Correo electrónico del usuario",
                        "type": "string",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "description": "Contraseña del usuario (Oculto)",
                        "type": "string"
                    },
                    "remember_me": {
                        "description": "Determina si el token durará más tiempo",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "title": "Login response",
                "description": "Login response",
                "properties": {
                    "access_token": {
                        "title": "access_token",
                        "description": "Access token",
                        "type": "string"
                    },
                    "token_type": {
                        "title": "token_type",
                        "description": "Tipo de token",
                        "type": "string"
                    },
                    "expires_at": {
                        "title": "expires_at",
                        "description": "Fecha de expiración",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "external_auth": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "http://example.swagger.io/oauth/dialog",
                        "scopes": {
                            "write:models": "modify models in your account",
                            "read:models": "read your models"
                        }
                    }
                }
            },
            "simple-api-key": {
                "type": "apiKey",
                "name": "simple-api-key",
                "in": "header"
            },
            "Bearer": {
                "type": "apiKey",
                "name": "bearer_auth",
                "in": "header"
            }
        }
    }
}